Refactored applicant service
