+++ src/figvalidationservice/src/main/java/ca/bc/gov/open/pssg/rsbc/dps/figvalidationservice/applicant/ApplicantController.java
@@ -2,13 +2,11 @@ package ca.bc.gov.open.pssg.rsbc.dps.figvalidationservice.applicant;
 
 import ca.bc.gov.open.ords.figcr.client.api.model.ValidateApplicantForSharingOrdsResponse;
 import ca.bc.gov.open.ords.figcr.client.api.model.ValidateApplicantPartyIdOrdsResponse;
+import ca.bc.gov.open.ords.figcr.client.api.model.ValidateApplicantServiceOrdsResponse;
 import ca.bc.gov.open.pssg.rsbc.dps.figvalidationservice.FigaroValidationServiceConstants;
-import ca.bc.gov.open.pssg.rsbc.dps.figvalidationservice.applicant.types.ValidateApplicantForSharingRequest;
-import ca.bc.gov.open.pssg.rsbc.dps.figvalidationservice.applicant.types.ValidateApplicantForSharingResponse;
+import ca.bc.gov.open.pssg.rsbc.dps.figvalidationservice.applicant.types.*;
 import ca.bc.gov.open.pssg.rsbc.dps.figvalidationservice.exception.FigaroValidationServiceException;
-import ca.bc.gov.open.pssg.rsbc.dps.figvalidationservice.applicant.types.ValidateApplicantPartyIdResponse;
-import ca.bc.gov.open.pssg.rsbc.dps.figvalidationservice.applicant.types.LocateMatchingApplicantsRequest;
-import ca.bc.gov.open.pssg.rsbc.dps.figvalidationservice.applicant.types.LocateMatchingApplicantsResponse;
+import ca.bc.gov.open.pssg.rsbc.dps.figvalidationservice.applicant.types.ValidateApplicantServiceResponse;
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiParam;
 import io.swagger.annotations.ApiResponse;
@@ -148,6 +146,35 @@ public class ApplicantController {
                             applAliasSurname3,
                             applAliasFirstName3,
                             applAliasSecondInitial3));
+
+    }
+
+    @RequestMapping(value = "/validateApplicantService",
+            produces = { MediaType.APPLICATION_XML_VALUE },
+            method = RequestMethod.GET)
+    @ApiOperation(value = "Validate Applicant Service", response = ValidateApplicantServiceResponse.class, tags={ APPLICANT_API })
+    @ApiResponses(value = {@ApiResponse(code = 200, message = "Successful operation", response = ValidateApplicantServiceResponse.class) })
+    public ValidateApplicantServiceResponse validateApplicantService(
+            @ApiParam(value = "orgPartyId", required = false) @RequestParam(value="orgPartyId", defaultValue="") String orgPartyId,
+            @ApiParam(value = "applPartyId", required = false) @RequestParam(value="applPartyId", defaultValue="") String applPartyId) throws FigaroValidationServiceException {
+
+
+        try {
+
+            ValidateApplicantServiceOrdsResponse _ordsResponse = applicantService
+                    .validateApplicantService(applPartyId, orgPartyId);
+
+            return new ValidateApplicantServiceResponse(_ordsResponse.getStatusMessage(),
+                    Integer.parseInt(_ordsResponse.getStatusCode()), _ordsResponse.getValidationResult()
+            );
+
+        } catch (FigaroValidationServiceException ex) {
+            logger.error("An exception occured in ValidateApplicantServiceResponse validateApplicantService() : " + ex.getMessage());
+            ex.printStackTrace();
+            return new ValidateApplicantServiceResponse(ex.getMessage(),
+                    FigaroValidationServiceConstants.VALIDATION_SERVICE_FAILURE_CD,
+                    FigaroValidationServiceConstants.VALIDATION_SERVICE_BOOLEAN_FALSE);
+        }
 
     }
