+++ src/figvalidationservice/src/main/java/ca/bc/gov/open/pssg/rsbc/dps/figvalidationservice/exception/FigaroValidationServiceException.java
@@ -16,10 +16,6 @@ public class FigaroValidationServiceException extends Exception {
 	 */
 	private static final long serialVersionUID = -5250088612082352045L;
 
-	public FigaroValidationServiceException(Throwable reason) {
-		super(reason);
-	}
-
 	public FigaroValidationServiceException(String message) {
 		super(message);
 	}
